+++ src/FcsWatch.Core/ProjectCoreCracker.fs
@@ -6,6 +6,7 @@ open System.IO
 open FSharp.Compiler.SourceCodeServices
 open System.Xml
 open Dotnet.ProjInfo.Workspace
+open Dotnet.ProjInfo.Workspace.FCS
 
 
 module MSBuildPrj = Dotnet.ProjInfo.Inspect
@@ -187,11 +188,16 @@ let getProjectOptionsFromProjectFile (file : string) = async {
         return result
 }
 
-let getProjectOptionsFromProjectFiles (projPaths : string list) = 
-    let config = 
-        let msbuildLocator = MSBuildLocator()
-        LoaderConfig.Default(msbuildLocator)
 
-    let loader = Loader.Create(config)
+let getProjectOptionsFromScript (checker: FSharpChecker) (scriptPath: string): FSharpProjectOptions = 
+    let msbuildLocator = MSBuildLocator()
 
-    loader.LoadProjects projPaths
+    let netFwInfo = 
+        NetFWInfoConfig.Default(msbuildLocator) |> NetFWInfo.Create
+
+    let fsxBinder = FsxBinder(netFwInfo, checker)
+
+    let tfm = netFwInfo.LatestVersion ()
+
+    fsxBinder.GetProjectOptionsFromScriptBy(tfm, scriptPath, File.ReadAllText scriptPath)
+    |> Async.RunSynchronously