+++ src/FcsWatch.Porta/Main.fs
@@ -7,14 +7,15 @@ module FcsWatch.Porta.Main
 
 open System
 open FSharp.Compiler.SourceCodeServices
-open FcsWatch.Core.CompilerTmpEmitter
 open FcsWatch.Core.Compiler
 open FcsWatch.Core
 open FcsWatch.Core.FcsWatcher
 
 open Extensions
+open Fake.IO
+open System.IO
 
-let runFcsWatcher (config: PortaConfig) entryProjectFile =
+let runFcsWatcher (config: PortaConfig) =
     logger <- Logger.create config.LoggerLevel
     let compiler =
         { new ICompiler<CheckResult> with 
@@ -22,21 +23,41 @@ let runFcsWatcher (config: PortaConfig) entryProjectFile =
                 let! result = CrackedFsproj.check config.UseEditFiles config.LiveCheckOnly checker crackedFsproj
                 return [|result|]
             }
-            member x.WarmCompile = false
+            member x.WarmCompile = true
             member x.Summary (_, _) = ""
         }
 
     let compilerTmpEmitter = CompilerTmpEmitter.create config
 
 
-    let config: FcsWatch.Core.Config = 
+    let coreConfig: FcsWatch.Core.Config = 
         { LoggerLevel = config.LoggerLevel
           WorkingDir = config.WorkingDir
           NoBuild = config.NoBuild
           UseEditFiles = config.UseEditFiles
-          ActionAfterStoppingWatcher = ignore }
+          ActionAfterStoppingWatcher = ignore
+          OtherFlags = config.OtherFlags }
 
-    let fcsWatcher = fcsWatcher compilerTmpEmitter compiler config entryProjectFile
+    let checker = FSharpChecker.Create(keepAssemblyContents = true)
 
-    Console.ReadLine() |> ignore
+    let fcsWatcher, _ = fcsWatcherAndCompilerTmpAgent checker compilerTmpEmitter compiler coreConfig config.Fsproj
 
+    if config.Watch then
+        Console.ReadLine() |> ignore
+    elif config.Eval then 
+        let makeSourceFileChanges fullPaths =
+            let makeFileChange fullPath : FileChange =
+                let fullPath = Path.getFullName fullPath
+
+                { FullPath = fullPath
+                  Name = Path.GetFileName fullPath
+                  Status = FileStatus.Changed }
+
+            FcsWatcherMsg.DetectSourceFileChanges <!> (List.map makeFileChange fullPaths)
+
+        let cache = fcsWatcher.PostAndReply FcsWatcherMsg.GetCache
+
+        fcsWatcher.PostAndReply (makeSourceFileChanges [cache.EntryCrackedFsproj.SourceFiles.[0]])
+
+        fcsWatcher.PostAndReply FcsWatcherMsg.WaitCompiled
+        |> ignore