+++ tests/FcsWatch.InteractionTests/InteractionTests.fs
@@ -27,11 +27,15 @@ let testSourceFile1InTestLib = datas </> @"TestLib1/Library.fs"
 
 let interactionTests =
     testList "interaction tests" [
-        testCase "base interaction test" <| fun _ ->
-            FcsWatch.Cli.Main.main [|"--project-file"; entryProjPath|]
+        testCase "manual reload test" <| fun _ ->
+            FcsWatch.Cli.Main.main [|"--project-file"; entryProjPath;"--logger-level"; "normal"; "--debuggable" |]
             |> ignore
 
-        ftestCase "auto reload test" <| fun _ ->
+        testCase "auto reload test" <| fun _ ->
             FcsWatch.Cli.Main.main [|"--project-file"; entryProjPath; "--no-build"|]
+            |> ignore
+
+        ftestCase "fslive cli test" <| fun _ ->
+            FsLive.Driver.main [| entryProjPath; "--watch"; "--loggerlevel:2"; "--send"|]
             |> ignore
     ]